server/web/requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
Flask==3.0.3
Flask-MySQLdb==2.0.0
fonttools==4.54.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mysqlclient==2.2.4
numpy==2.0.2
packaging==24.1
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
urllib3==2.2.3
Werkzeug==3.0.4
zipp==3.20.2
pwntools